OSS_Library/pyproject.toml
[project]
name = "jeonju-library-map"
version = "0.1.0"
description = "Streamlit + FastAPI + SQLite3 starter for a Jeonju library availability map app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "requests>=2.32.0",
]

[dependency-groups]
backend = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "pydantic-settings>=2.8.0",
    "PyJWT>=2.10.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.34.0",
]
frontend = [
    "pydeck>=0.9.1",
    "streamlit>=1.44.0",
]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
